+++ src/nextjournal/clerk/builder.clj
@@ -2,12 +2,14 @@
   "Clerk's Static App Builder."
   (:require [babashka.fs :as fs]
             [clojure.java.browse :as browse]
+            [clojure.set :as set]
             [clojure.string :as str]
             [nextjournal.clerk.analyzer :as analyzer]
+            [nextjournal.clerk.builder-ui :as builder-ui]
             [nextjournal.clerk.eval :as eval]
             [nextjournal.clerk.parser :as parser]
-            [nextjournal.clerk.view :as view]))
-
+            [nextjournal.clerk.view :as view]
+            [nextjournal.clerk.webserver :as webserver]))
 
 (def clerk-docs
   (into ["CHANGELOG.md"
@@ -56,7 +58,7 @@
 #_(strip-index "hello_index.cljc")
 
 (defn ->html-extension [path]
-  (str/replace path #"\.(clj|cljc|md)$" ".html"))
+  (str/replace path #"\.(cljc?|md)$" ".html"))
 
 #_(->html-extension "hello.clj")
 
@@ -66,77 +68,124 @@
   [path]
   (str/replace path fs/file-separator "/"))
 
-(defn process-build-opts [{:as opts :keys [paths]}]
-  (merge {:out-path (str "public" fs/file-separator "build")
-          :bundle? true
-          :browse? true}
-         opts))
+(defn describe-event [{:as event :keys [stage state duration doc]}]
+  (let [format-duration (partial format "%.3fms")
+        duration (some-> duration format-duration)]
+    (case stage
+      :init (str "👷🏼 Clerk is building " (count state) " notebooks…\n🧐 Parsing… ")
+      :parsed (str "Done in " duration ". ✅\n🔬 Analyzing… ")
+      (:built :analyzed :done) (str "Done in " duration ". ✅\n")
+      :building (str "🔨 Building \"" (:file doc) "\"… ")
+      :downloading-cache (str "⏬ Downloading distributed cache… ")
+      :uploading-cache (str "⏫ Uploading distributed cache… ")
+      :finished (str "📦 Static app bundle created in " duration ". Total build time was " (-> event :total-duration format-duration) ".\n"))))
+
+(defn stdout-reporter [build-event]
+  (doto (describe-event build-event)
+    (print)
+    (do (flush))))
+
+(defn build-ui-reporter [{:as build-event :keys [stage]}]
+  (when (= stage :init)
+    (builder-ui/reset-build-state!)
+    ((resolve 'nextjournal.clerk/show!) (clojure.java.io/resource "nextjournal/clerk/builder_ui.clj"))
+    (when-let [{:keys [port]} (and (get-in build-event [:build-opts :browse]) @webserver/!server)]
+      (browse/browse-url (str "http://localhost:" port))))
+  (stdout-reporter build-event)
+  (builder-ui/add-build-event! build-event)
+  (binding [*out* (java.io.StringWriter.)]
+    ((resolve 'nextjournal.clerk/recompute!))))
+
+(def default-out-path
+  (str "public" fs/file-separator "build"))
+
+(defn ^:private migrate-deprecated-opts [opts]
+  (set/rename-keys opts {:bundle? :bundle
+                         :browse? :browse}))
+
+(defn process-build-opts [{:as opts :keys [paths index]}]
+  (merge {:out-path default-out-path
+          :bundle false
+          :browse false
+          :report-fn (if @webserver/!server build-ui-reporter stdout-reporter)}
+         (cond-> (migrate-deprecated-opts opts)
+           index (assoc :index (str index)))))
+
+#_(process-build-opts {:index 'book.clj})
 
 (defn write-static-app!
-  "Creates a static html app of the seq of `docs`. Customizable with an `opts` map with keys:
-
-  - `:paths` a vector of relative paths to notebooks to include in the build
-  - `:bundle?` builds a single page app versus a folder with an html page for each notebook (defaults to `true`)
-  - `:out-path` a relative path to a folder to contain the static pages (defaults to `\"public/build\"`)
-  - `:git/sha`, `:git/url` when both present, each page displays a link to `(str url \"blob\" sha path-to-notebook)`"
   [opts docs]
-  (let [{:as opts :keys [bundle? out-path browse?]} (process-build-opts opts)
+  (let [{:as opts :keys [bundle out-path browse index]} (process-build-opts opts)
         paths (mapv :file docs)
         path->doc (into {} (map (juxt :file :viewer)) docs)
-        path->url (into {} (map (juxt identity #(cond-> (strip-index %) (not bundle?) ->html-extension))) paths)
-        static-app-opts (assoc opts :bundle? bundle? :path->doc path->doc :paths (vec (keys path->doc)) :path->url path->url)
+        path->url (into {} (map (juxt identity #(cond-> (strip-index %) (not bundle) ->html-extension))) paths)
+        static-app-opts (assoc opts :bundle bundle :path->doc path->doc :paths (vec (keys path->doc)) :path->url path->url)
         index-html (str out-path fs/file-separator "index.html")]
     (when-not (fs/exists? (fs/parent index-html))
       (fs/create-dirs (fs/parent index-html)))
-    (if bundle?
+    (if bundle
       (spit index-html (view/->static-app static-app-opts))
-      (do (when-not (contains? (-> path->url vals set) "") ;; no user-defined index page
+      (do (when-not (contains? (-> path->url vals set) (->html-extension (str index))) ;; no user-defined index page
             (spit index-html (view/->static-app (dissoc static-app-opts :path->doc))))
           (doseq [[path doc] path->doc]
-            (let [out-html (str out-path fs/file-separator (str/replace path #"(.cljc?|.md)" ".html"))]
+            (let [path-with-index-mapped (if index ({index "index.clj"} path path) path)
+                  out-html (str out-path fs/file-separator (->html-extension path-with-index-mapped))]
               (fs/create-dirs (fs/parent out-html))
               (spit out-html (view/->static-app (assoc static-app-opts :path->doc (hash-map path doc) :current-path path)))))))
-    (when browse?
+    (when browse
       (browse/browse-url (-> index-html fs/absolutize .toString path-to-url-canonicalize)))
-    docs))
+    {:docs docs
+     :index-html index-html
+     :build-href (if (and @webserver/!server (= out-path default-out-path)) "/build" index-html)}))
 
-(defn stdout-reporter [{:as event :keys [stage state duration doc]}]
-  (let [format-duration (partial format "%.3fms")
-        duration (some-> duration format-duration)]
-    (print (case stage
-             :init (str "👷🏼 Clerk is building " (count state) " notebooks…\n🧐 Parsing… ")
-             :parsed (str "Done in " duration ". ✅\n🔬 Analyzing… ")
-             (:built :analyzed :done) (str "Done in " duration ". ✅\n")
-             :building (str "🔨 Building \"" (:file doc) "\"… ")
-             :downloading-cache (str "⏬ Downloading distributed cache… ")
-             :uploading-cache (str "⏫ Uploading distributed cache… ")
-             :finished (str "📦 Static app bundle created in " duration ". Total build time was " (-> event :total-duration format-duration) ".\n"))))
-  (flush))
+(defn ^:private maybe-add-index [{:as build-opts :keys [paths paths-fn index]} resolved-paths]
+  (when (and index (or (not (string? index)) (not (fs/exists? index))))
+    (throw (ex-info "`:index` must be string and point to existing file" {:index index})))
+  (cond-> resolved-paths
+    (and index (not (contains? (set resolved-paths) index)))
+    (conj index)))
 
-(defn expand-paths [paths]
-  (->> (if (symbol? paths)
-         (let [resolved (-> paths requiring-resolve deref)]
-           (cond-> resolved
-             (fn? resolved) (apply [])))
-         paths)
+#_(maybe-add-index {:index "book.clj"} nil)
+
+(defn expand-paths [{:as build-opts :keys [paths paths-fn index]}]
+  (when (and paths paths-fn)
+    (binding [*out* *err*]
+      (println "[info] both `:paths` and `:paths-fn` are set, `:paths` will take precendence.")))
+  (when (not (or paths paths-fn index))
+    (throw (ex-info "must set either `:paths`, `:paths-fn` or `:index`." {:build-opts build-opts})))
+  (->> (cond paths (if (sequential? paths)
+                     paths
+                     (throw (ex-info "`:paths` must be sequential" {:paths paths})))
+             paths-fn (if (qualified-symbol? paths-fn)
+                        (try
+                          (if-let [resolved-var  (requiring-resolve paths-fn)]
+                            (let [resolved-paths (cond-> @resolved-var
+                                                   (fn? @resolved-var) (apply []))]
+                              (when-not (sequential? resolved-paths)
+                                (throw (ex-info (str "#'" paths-fn " must be sequential.") {:paths-fn paths-fn :resolved-paths resolved-paths})))
+                              resolved-paths)
+                            (throw (ex-info (str "#'" paths-fn " cannot be resolved.") {:paths-fn paths-fn}))))
+                        (throw (ex-info "`:path-fn` must be a qualified symbol pointing at an existing var." {:paths-fn paths-fn}))))
+       (maybe-add-index build-opts)
        (mapcat (partial fs/glob "."))
        (filter (complement fs/directory?))
        (mapv (comp str fs/file))))
 
-#_(expand-paths ["notebooks/di*.clj"])
-#_(expand-paths `clerk-docs)
+#_(expand-paths {:paths ["notebooks/di*.clj"]})
+#_(expand-paths {:paths ['notebooks/rule_30.clj]})
+#_(expand-paths {:paths-fn `clerk-docs})
+#_(expand-paths {:paths-fn `clerk-docs-2})
 #_(do (defn my-paths [] ["notebooks/h*.clj"])
-      (expand-paths `my-paths))
-#_(expand-paths ["notebooks/viewers**"])
+      (expand-paths {:paths-fn `my-paths}))
+#_(expand-paths {:paths ["notebooks/viewers**"]})
 
 (defn build-static-app! [opts]
-  (let [{:as opts :keys [expanded-paths paths download-cache-fn upload-cache-fn bundle?]} (assoc (process-build-opts opts) :expanded-paths (-> opts :paths expand-paths))
+  (let [{:as opts :keys [expanded-paths paths download-cache-fn upload-cache-fn bundle report-fn]} (assoc (process-build-opts opts) :expanded-paths (expand-paths opts))
         _ (when (empty? expanded-paths)
             (throw (ex-info "nothing to build" {:expanded-paths expanded-paths :paths paths})))
         start (System/nanoTime)
-        report-fn stdout-reporter
         state (mapv #(hash-map :file %) expanded-paths)
-        _ (report-fn {:stage :init :state state})
+        _ (report-fn {:stage :init :state state :build-opts opts})
         {state :result duration :time-ms} (eval/time-ms (mapv (comp (partial parser/parse-file {:doc? true}) :file) state))
         _ (report-fn {:stage :parsed :state state :duration duration})
         {state :result duration :time-ms} (eval/time-ms (mapv (comp analyzer/hash
@@ -146,21 +195,22 @@
             (report-fn {:stage :downloading-cache})
             (let [{duration :time-ms} (eval/time-ms (download-cache-fn state))]
               (report-fn {:stage :done :duration duration})))
-        state (mapv (fn [doc]
-                      (report-fn {:stage :building :doc doc})
-                      (let [{doc+viewer :result duration :time-ms} (eval/time-ms
-                                                                    (let [doc (eval/eval-analyzed-doc doc)]
-                                                                      (assoc doc :viewer (view/doc->viewer (assoc opts :inline-results? true) doc))))]
-                        (report-fn {:stage :built :doc doc+viewer :duration duration})
-                        doc+viewer)) state)
-        {state :result duration :time-ms} (eval/time-ms (write-static-app! opts state))]
+        docs (mapv (fn [doc idx]
+                     (report-fn {:stage :building :doc doc :idx idx})
+                     (let [{doc+viewer :result duration :time-ms} (eval/time-ms
+                                                                   (let [doc (eval/eval-analyzed-doc doc)]
+                                                                     (assoc doc :viewer (view/doc->viewer (assoc opts :inline-results? true) doc))))]
+                       (report-fn {:stage :built :doc doc+viewer :duration duration :idx idx})
+                       doc+viewer)) state (range))
+        {state :result duration :time-ms} (eval/time-ms (write-static-app! opts docs))]
     (when upload-cache-fn
       (report-fn {:stage :uploading-cache})
       (let [{duration :time-ms} (eval/time-ms (upload-cache-fn state))]
         (report-fn {:stage :done :duration duration})))
     (report-fn {:stage :finished :state state :duration duration :total-duration (eval/elapsed-ms start)})))
 
-#_(build-static-app! {:paths (take 5 clerk-docs)})
-#_(build-static-app! {:paths ["index.clj" "notebooks/rule_30.clj" "notebooks/viewer_api.md"] :bundle? true})
-#_(build-static-app! {:paths ["index.clj" "notebooks/rule_30.clj" "notebooks/viewer_api.md"] :bundle? false})
+#_(build-static-app! {:paths (take 15 clerk-docs)})
+#_(build-static-app! {:paths ["index.clj" "notebooks/rule_30.clj" "notebooks/viewer_api.md"] :bundle true})
+#_(build-static-app! {:paths ["index.clj" "notebooks/rule_30.clj" "notebooks/markdown.md"] :bundle false :browse false})
 #_(build-static-app! {:paths ["notebooks/viewers/**"]})
+#_(build-static-app! {:index "notebooks/rule_30.clj"  :git/sha "bd85a3de12d34a0622eb5b94d82c9e73b95412d1" :git/url "https://github.com/nextjournal/clerk"})